first

+++ database/migrations/2017_06_03_180426_create_sp_video_types_table.php
@@ -17,8 +17,8 @@ class CreateSpVideoTypesTable extends Migration
             $table->increments('id');
             $table->string('name')->default('')->index()->comment("类型名称");
             $table->string('remark')->default('')->comment("备注");
-            $table->tinyInteger('sort')->default(0)->comment("排序");
-            $table->tinyInteger('status')->default(1)->comment("状态：1正常");
+            $table->integer('sort')->default(0)->comment("排序");
+            $table->integer('status')->default(1)->comment("状态：1正常");
         });
     }
